cvsms-app-api/ApiFlaskApp/requirements.txt
beautifulsoup4==4.3.2
cassandra-driver==3.7.1
cql==1.4.0
Flask==0.10.1
Flask-Cors==3.0.2
futures==3.0.5
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pika==0.10.0
requests==2.4.1
six==1.8.0
thrift==0.9.3
Werkzeug==0.9.6
